backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "livekit>=1.1.6",
    "livekit-api>=1.1.0",
    "livekit-protocol>=1.1.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
